+++ server/config/config.go
@@ -77,9 +77,6 @@ const (
 	// DefaultJSONRPCAllowInsecureUnlock is true
 	DefaultJSONRPCAllowInsecureUnlock bool = true
 
-	// DefaultFilterCap is the default cap for total number of filters that can be created
-	DefaultFilterCap int32 = 200
-
 	// DefaultFilterTimeout defines when an idle filter expires.
 	DefaultFilterTimeout = 5 * time.Minute
 
@@ -262,8 +259,6 @@ type JSONRPCConfig struct {
 	EVMTimeout time.Duration `mapstructure:"evm-timeout"`
 	// TxFeeCap is the global tx-fee cap for send transaction
 	TxFeeCap float64 `mapstructure:"txfee-cap"`
-	// FilterCap is the global cap for total number of filters that can be created.
-	FilterCap int32 `mapstructure:"filter-cap"`
 	// FilterTimeout defines when an idle filter expires.
 	FilterTimeout time.Duration `mapstructure:"filter-timeout"`
 	// FilterCleanupInterval defines how often expired filters are cleaned up.
@@ -366,7 +361,6 @@ func DefaultJSONRPCConfig() *JSONRPCConfig {
 		AllowInsecureUnlock:   DefaultJSONRPCAllowInsecureUnlock,
 		EVMTimeout:            DefaultEVMTimeout,
 		TxFeeCap:              DefaultTxFeeCap,
-		FilterCap:             DefaultFilterCap,
 		FilterTimeout:         DefaultFilterTimeout,
 		FilterCleanupInterval: DefaultFilterCleanupInterval,
 		FeeHistoryCap:         DefaultFeeHistoryCap,
@@ -390,10 +384,6 @@ func DefaultJSONRPCConfig() *JSONRPCConfig {
 func (c JSONRPCConfig) Validate() error {
 	if c.Enable && len(c.API) == 0 {
 		return errors.New("cannot enable JSON-RPC without defining any API namespace")
-	}
-
-	if c.FilterCap < 0 {
-		return errors.New("JSON-RPC filter-cap cannot be negative")
 	}
 
 	if c.FilterTimeout < 0 {